fix: webworker build error

+++ packages/cysync-core/workers/build.sh
@@ -1,12 +1,14 @@
+BUFFER_FILE_PATH='../scripts/dependencies/dist/buffer.js'
 FOLDERPATH='../src/generated/workers'
 FILEPATH='../src/generated/workers/graph.js'
 
 cd workers
 
-rm -rf ${FOLDERPATH}/*
+rm -rf ${FOLDERPATH:?}/*
 mkdir -p ${FOLDERPATH}
 
 pnpm build
 
 echo "/* eslint-disable */" > ${FILEPATH}
-cp parcel/index.js ${FILEPATH}
+cat ${BUFFER_FILE_PATH} >> ${FILEPATH}
+cat parcel/index.js >> ${FILEPATH}
